requirements.txt
attrs==21.2.0
black==21.9b0
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
#regex==2021.10.23
six==1.16.0
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
google-cloud-pubsub==2.11.0
google-cloud-storage==2.2.1
google-cloud-bigquery==2.32.0
